osx: 64bit: fix more various variable sizes

+++ Source/OS9exec_core/modstuff.c
@@ -128,7 +128,7 @@
 #endif
 
 /* additional memory for all processes */
-ulong memplusall;
+uint32_t memplusall;
 
 /* Memory modules */
 /* ============== */
